fix(ci): pass all quality gates — clippy, fuzz smoke, flutter test
- Fix 35+ clippy errors: needless_range_loop, doc indentation,
  field_reassign_with_default, new_without_default, too_many_arguments,
  unnecessary mut, infinite loop → while-let, identity match arm
- Remove `working-directory: fuzz` from all three fuzz smoke steps;
  cargo-fuzz must run from repo root to locate fuzz/Cargo.toml
- Remove assert(bridge.isAvailable) from app.dart initState — every
  bridge method already guards with `if (!isAvailable) return …`; the
  assert contradicts allowMissing:true used in flutter widget tests

+++ backend/service/file_ops.rs
@@ -701,7 +701,7 @@ fn storage_capacity_bytes(path: &str) -> Option<u64> {
     }
 
     let stats = unsafe { stats.assume_init() };
-    Some((stats.f_blocks as u64).saturating_mul(stats.f_frsize as u64))
+    Some(stats.f_blocks.saturating_mul(stats.f_frsize))
 }
 
 #[cfg(not(unix))]
